requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
lxml==5.2.2
python-slugify==8.0.4
requests==2.32.3
soupsieve==2.5
text-unidecode==1.3
urllib3==2.2.2
